Optimize architecture
