+++ rekcurd_dashboard/console_scripts/__init__.py
@@ -4,36 +4,13 @@
 import argparse
 
 from rekcurd_dashboard import _version
-from rekcurd_dashboard.server import create_app
 
-
-def server_handler(args):
-    app = create_app(args.settings, args.logger)
-    app.run(host=args.host, port=args.port, threaded=True)
-
-
-def db_handler(args):
-    import sys
-    from flask_script import Manager
-    from flask_migrate import Migrate, MigrateCommand
-    from rekcurd_dashboard.models import db
-
-    tmp = sys.argv[1:]
-    sys.argv = [sys.argv[0]]
-    for i, t in enumerate(tmp):
-        if t == 'db':
-            sys.argv.append(t)
-            sys.argv.append(tmp[i+1])
-            break
-    app = create_app(args.settings, args.logger)
-    migrate = Migrate(app, db)
-    manager = Manager(app)
-    manager.add_command('db', MigrateCommand)
-    manager.run()
+from .server_handler import server_handler
+from .db_handler import db_handler
 
 
 def create_parser():
-    parser = argparse.ArgumentParser(description='rekcurdui command')
+    parser = argparse.ArgumentParser(description='rekcurd_dashboard command')
     parser.add_argument(
         '--version', '-v', action='version', version=_version.__version__)
     parser.add_argument(
@@ -44,7 +21,7 @@ def create_parser():
 
     # server
     parser_server = subparsers.add_parser(
-        'server', help='see `rekcurdui server -h`')
+        'server', help='see `rekcurd_dashboard server -h`')
     parser_server.add_argument(
         '-H', '--host', required=False, help='host', default='0.0.0.0')
     parser_server.add_argument(
@@ -53,7 +30,7 @@ def create_parser():
 
     # db
     parser_db = subparsers.add_parser(
-        'db', help='see `rekcurdui db -h`')
+        'db', help='see `rekcurd_dashboard db -h`')
     parser_db.add_argument(
         'type', choices=['init', 'revision', 'migrate', 'edit', 'merge',
                          'upgrade', 'downgrade', 'show', 'history', 'heads',
