+++ rekcurd_dashboard/console_scripts/server_handler.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from rekcurd_dashboard.core import create_app
+
+
+def server_handler(args):
+    app = create_app(args.settings, args.logger)
+    app.run(host=args.host, port=args.port, threaded=True)
