+++ rekcurd_dashboard/core/__init__.py
@@ -1,0 +1,2 @@
+from .rekcurd_dashboard_client import RekcurdDashboardClient
+from .create_app import create_app
