+++ tox.ini
@@ -24,7 +24,7 @@ commands =
 [testenv:coverage]
 commands =
    python -V
-   {toxinidir}/scripts/kube-init.sh nosetests --with-coverage --cover-package=rekcurd_dashboard --cover-tests
+   {toxinidir}/scripts/kube-init.sh nosetests --with-coverage --cover-package=rekcurd_dashboard.apis,rekcurd_dashboard.auth,rekcurd_dashboard.core,rekcurd_dashboard.logger,rekcurd_dashboard.models,rekcurd_dashboard.utils --cover-tests
 
 [testenv:codecov]
 commands =
